extracted a Runtime method to a new class
