+++ mRemoteV1/mRemoteV1.csproj
@@ -645,6 +645,7 @@
     <Compile Include="UI\Menu\ViewMenu.cs">
       <SubType>Component</SubType>
     </Compile>
+    <Compile Include="UI\Panel\PanelAdder.cs" />
     <Compile Include="UI\TaskDialog\CommandButton.cs">
       <SubType>Component</SubType>
     </Compile>
